+++ billing/contracts/charges/Charge.sol
@@ -2,7 +2,7 @@ pragma solidity ^0.4.11;
 
 contract Charge {
   
-  function getPrice(uint _callID, address _from) returns (uint);
+  function getPrice(uint256 _callID, address _from) returns (uint256);
 
   function resetToken(address _from);
